// src/api/invitation/types.ts
/* eslint-disable */
// tslint:disable
/**
 * Yeying API
 * 
 *
 * OpenAPI spec version: 1.0
 * Contact: yeying.community@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface InvitationApi {
	invitationCreate: (request: Api.InvitationCreateInvitationRequest) => Promise<InvitationCreateResponse>
	invitationDetail: (request: Api.InvitationInvitationDetailRequest) => Promise<InvitationDetailResponse>
	invitationSearch: (request: Api.InvitationSearchInvitationRequest) => Promise<InvitationSearchResponse>
}

export type InvitationCreateResponse = InvitationCreate200Response | InvitationCreateDefaultResponse

export interface InvitationCreate200Response {
	status: 200
	body: Api.InvitationCreateInvitationResponse
	headers?: never
}

export interface InvitationCreateDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}

export type InvitationDetailResponse = InvitationDetail200Response | InvitationDetailDefaultResponse

export interface InvitationDetail200Response {
	status: 200
	body: Api.InvitationInvitationDetailResponse
	headers?: never
}

export interface InvitationDetailDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}

export type InvitationSearchResponse = InvitationSearch200Response | InvitationSearchDefaultResponse

export interface InvitationSearch200Response {
	status: 200
	body: Api.InvitationSearchInvitationResponse
	headers?: never
}

export interface InvitationSearchDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}